Added submission controller method to get submissions for logged in provider id
Decided to add in Spring Security to that provider id is passed from security principal coming in a token. At present this is all faked as we don't have real auth yet. CSRF and HTTP authentication are turned off. We need to think about whether the whole API is in the context of the logged in user or whether any parts are machine based, e.g. from the bulk worker service.